requirements.txt
pydantic_settings
pydantic
uvicorn
fastapi
sqlalchemy
greenlet
pyjwt
cryptography
pydantic[email]
bcrypt